feat(gh-16): update for the 1.0 release
