+++ activity/PlayerActivity.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using System.Text.Json;
-using System.Text.Json.Serialization;
 
 namespace v_rising_discord_bot_companion.activity;
 
@@ -9,20 +7,3 @@ public readonly record struct PlayerActivity(
     string PlayerName,
     DateTime Occurred
 );
-
-[JsonConverter(typeof(ActivityTypeConverter))]
-public enum ActivityType {
-    CONNECTED,
-    DISCONNECTED
-}
-
-public class ActivityTypeConverter : JsonConverter<ActivityType> {
-
-    public override ActivityType Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options) {
-        throw new NotImplementedException();
-    }
-
-    public override void Write(Utf8JsonWriter writer, ActivityType value, JsonSerializerOptions options) {
-        writer.WriteStringValue(Enum.GetName(typeof(ActivityType), value));
-    }
-}
