+++ activity/ServerBootstrapSystemPatches.cs
@@ -13,7 +13,7 @@ namespace v_rising_discord_bot_companion.activity;
 [HarmonyPatch(typeof(ServerBootstrapSystem))]
 public class ServerBootstrapSystemPatches {
 
-    private static readonly List<PlayerActivity> _playerActivities = new();
+    private readonly static List<PlayerActivity> _playerActivities = new();
 
     public static List<PlayerActivity> getPlayerActivities() {
         removeExpiredPlayerActivities();
