+++ game/VCharacter.cs
@@ -1,0 +1,26 @@
+﻿using Bloodstone.API;
+using ProjectM;
+using Unity.Entities;
+
+namespace v_rising_discord_bot_companion.game;
+
+public readonly record struct VCharacter(
+    PlayerCharacter Character,
+    Entity CharacterEntity
+) {
+
+    public static VCharacter from(VUser vUser) {
+        return from(vUser.User.LocalCharacter._Entity);
+    }
+
+    public static VCharacter from(Entity characterEntity) {
+        return new VCharacter(
+            Character: VWorld.Server.EntityManager.GetComponentData<PlayerCharacter>(characterEntity),
+            CharacterEntity: characterEntity
+        );
+    }
+
+    public int getGearLevel() {
+        return (int) VWorld.Server.EntityManager.GetComponentData<Equipment>(CharacterEntity).GetFullLevel();
+    }
+}
