+++ game/VPlayer.cs
@@ -1,7 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using Bloodstone.API;
-using ProjectM;
 using ProjectM.Network;
 using Unity.Collections;
 using Unity.Entities;
@@ -41,40 +40,5 @@ public readonly record struct VPlayer(
 
     public bool HasCharacter() {
         return VCharacter != null;
-    }
-}
-
-public readonly record struct VUser(
-    User User,
-    Entity UserEntity
-) {
-
-    public static VUser from(Entity userEntity) {
-        var user = VWorld.Server.EntityManager.GetComponentData<User>(userEntity);
-        return new VUser(
-            User: user,
-            UserEntity: userEntity
-        );
-    }
-}
-
-public readonly record struct VCharacter(
-    PlayerCharacter Character,
-    Entity CharacterEntity
-) {
-
-    public static VCharacter from(VUser vUser) {
-        return from(vUser.User.LocalCharacter._Entity);
-    }
-
-    public static VCharacter from(Entity characterEntity) {
-        return new VCharacter(
-            Character: VWorld.Server.EntityManager.GetComponentData<PlayerCharacter>(characterEntity),
-            CharacterEntity: characterEntity
-        );
-    }
-
-    public int getGearLevel() {
-        return (int) VWorld.Server.EntityManager.GetComponentData<Equipment>(CharacterEntity).GetFullLevel();
     }
 }
