+++ killfeed/Player.cs
@@ -1,0 +1,6 @@
+﻿namespace v_rising_discord_bot_companion.killfeed;
+
+public readonly record struct Player(
+    string Name,
+    int GearLevel
+);
