+++ query/AsyncQuery.cs
@@ -2,10 +2,6 @@
 
 namespace v_rising_discord_bot_companion.query;
 
-public interface Query {
-    public void Invoke();
-}
-
 public class AsyncQuery<T> : Query {
 
     public Status Status { get; private set; }
