+++ query/Query.cs
@@ -1,0 +1,6 @@
+﻿namespace v_rising_discord_bot_companion.query;
+
+public interface Query {
+
+    public void Invoke();
+}
